A LOT! New Fonts, New Header
A new header has been added with a border and image to welcome the page.
Added many new fonts to webpage. There is a blank js script linked and
will be added in future (currently moved on a branch). Dont worry about
console error as the file will be added and will not make any difference
to code.